+++ B-E/src/main/java/com/example/job_portal/jobportal/Service/JobService.java
@@ -1,17 +1,11 @@
 package com.example.job_portal.jobportal.Service;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.example.job_portal.jobportal.Repository.JobRepository;
-import com.example.job_portal.jobportal.module.Companies;
 import com.example.job_portal.jobportal.module.Jobs;
 
 @Service
@@ -30,15 +24,15 @@ public class JobService {
         }
 
 
-    public void deleteJobs(int id) {
+    public void deleteJobs(Long id) {
         jobRepository.deleteById(id);
     }
 
-    public Jobs getJob(Integer id) {
+    public Jobs getJob(Long id) {
         return jobRepository.findById(id).orElse(null); 
     }
 
-     public String editJob(int id,Jobs jobDetails) {
+     public String editJob(Long id,Jobs jobDetails) {
         jobDetails.setId(id);
         jobRepository.save(jobDetails);
         return "Company edited successfully!";
